services/driver/requirements.txt
# FastAPI server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# SQLAlchemy and MySQL
sqlalchemy==2.0.30
pymysql==1.1.0

# Alembic for migrations
alembic==1.13.1

# Environment variable support
python-dotenv==1.0.1

cryptography 

pydantic==1.10.13